refactor(supplier): change dto's folder
