--- services/inventory/src/main/java/com/petek/inventoryService/dto/SupplierResponse.java
@@ -1,21 +1,0 @@
-package com.petek.inventoryService.dto;
-
-import java.time.Instant;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class SupplierResponse {
-    private Long supplierId;
-    private String supplierName;
-    private String email;
-    private String phone;
-    private String city;
-    private Instant createdAt;
-}
